requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
orjson==3.6.7
psycopg2==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
SCons==4.3.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.9
uvicorn==0.15.0
watchgod==0.8
websockets==10.2
